Fix flaky go-aks test 'SP not found' error with a simple sleep (#3445)
Although Azure returns success on creation of the Service Principal, we
would often see _"Searching service principal failed. Details: service
principal is not found"_ errors without this `sleep`. I ran if five
times in a row and it passed each time - hopefully good enough to merge
it and give it a try.

Fixes #3365