Azure-in-Azure integration test that runs Pulumi in Azure to test managed identity auth (#3449)
This PR introduces an integration test that consists of two Pulumi
programs, the outer and the inner one. The outer one is run by the test
suite via ProgramTest as usual. It creates an Azure VM, copies the inner
program on that VM, and runs it there.

This allows testing managed identity authentication, which requires
running on Azure VMs or similar services. The inner program running on
Azure is configured with `ARM_USE_MSI=true` and has no other means to
authenticate, so it would fail without working managed identity auth.

What the inner program actually provisions doesn't matter because we
only want to test auth. Experimentation showed that it does need to
provision something; preview alone is not enough because even without
managed identity enabled, the token endpoint on the VM is up and gives
out non-working tokens.

For this provisioning to work, the VM's identity needs to have a role
assigned that allows it to provision and remove resources. We use the
Azure built-in Contributor role. In order to minimize attack surface and
avoid a potential security issue by leaking a VM that has access to the
whole subscription, the scope of the role assignment is limited to the
temporary resource group created by the outer program. For the inner
program to be able to use that resource group, the outer program passes
its resource id to the inner program via pulumi config and the inner
program references it via
[get](https://www.pulumi.com/docs/languages-sdks/yaml/yaml-language-reference/#resources).