Expect refresh changes to WebApp.SiteConfig because it's modified by other resources (#3482)
Follow-up PR to #3464. That PR added a proper Read of WebApp.SiteConfig
to Reads of WebApp. That works fine as long the web app is only modified
through the WebApp resource. However, the Azure spec and the provider
also expose dedicated resources for some web app aspects, like
[native.web.WebAppAuthSettings](https://www.pulumi.com/registry/packages/azure-native/api-docs/web/webappauthsettings/)
and others.

When those resources are used, we get the following sequence:
1. WebApp is created with `siteConfig.foo = A`.
2. WebAppSomething is created and sets `siteConfig.foo = B`.
3. A Refresh of the WebApp shows a change in `siteConfig.foo`.

It's even worse if the original WebApp resource didn't specify
siteConfig at all. Then, as the property is missing, the provider's diff
inserts the full siteConfig from the Read, causing the diff to show the
whole siteConfig.

This problem is an instance of the known problem of sub-resources, see
#2755. That solution might be applicable here as well, but would need to
be generalized first, as it expects a list-type property of
sub-resources currently, whereas siteConfig is a singleton.

[Successful full test
run](https://github.com/pulumi/pulumi-azure-native/actions/runs/10210742989/job/28251434378),
except expected OIDC failures.