Add an exception for broken VMSS.diskController property type: should be string (#3484)
Special case for
[VirtualMachineScaleSet.diskControllerType](https://www.pulumi.com/registry/packages/azure-native/api-docs/compute/virtualmachinescaleset/#diskcontrollertype_go).
This property is erroneously defined as a string _and_ an array in the
spec. Primitive string is correct. See #2639, upstream issue
Azure/azure-rest-api-specs#25202. Fixes #2639.

Note: I tried going with the primitive type over the array for all such
properties but it caused a large diff in the azurestackhci SDK, so I
opted not to do that.

There's no test since there's no test coverage for the whole complex
`genTypeSpec` method. That would have made it necessary to refactor and
invent a proper testing harness which was beyond the scope of this quick
fix. The SDK output is pretty obvious to review. I also tested it in a
local program.