Extend the custom handling of SiteConfig and deleteEmptyServerFarm to WebAppSlot (#3495)
This PR extends the changes of both #3464 and #3004 from the WebApp to
the WebAppSlot resource. Web apps and web apps slot are virtually
identical and share much of their Azure spec as well, therefore, the
fixes apply to them as well.

The only net-new code is in adjusting the tests and some refactoring,
since both resources share the actual CRUD implementation.

The integration test is extended to create and verify both resource
types.