Upgrade specs & remove manual portal.Dashboard implementation (#3500)
Fixes #3489 

Revert #3123 which hard coded types for the dashboard resource. The
specifications have been updated and improved which means the mixins
caused the schema to be invalid. Our manual mixins are no longer needed
as the specification appears to include the same level of detail. The
structure of the types have changed slightly which will be a minor
breaking change for existing usage which will require users to update
their code.

## Migrating existing dashboard code

1. The top-level `lenses` and `metadata` properties are now wrapped
inside a `properties` object.
2. The only supported "Part Type" is
`Extension/HubsExtension/PartType/MarkdownPart`.
3. The `properties.lenses.parts.metadata.settings` property has now got
concrete types instead of being a dictionary.
4. The names of some sub-types have changed.