Remove erroneous useOids and useMsi defaults that interfere with env variables (#3525)
This PR is a follow-up to #3180, which was correct but incomplete, and
probably resolves #4779.

Tested locally.

Copying the description of 3180, slightly adjusted:

### Scenario
- The environment has `ARM_USE_OIDC=true`.
- An explicit provider is created without setting the `useOidc`
parameter.

### Expectation
OIDC authentication is enabled and being used (provided the other
necessary configurations are set), due to `ARM_USE_OIDC`.

### Actual
OIDC authentication is not used. 

Verbose logs show that when the explicit provider runs `getAuthConfig`,
the `useOidc` Pulumi setting is always `false`, having been set based on
the default value. The environment variable is never checked.

All of the above is also true for `ARM_USE_MSI`/`useMsi`.