Replace Autorest with azcore (#3536)
This PR introduces a new implementation of `AzureClient` that's based on
[azcore](https://pkg.go.dev/github.com/Azure/azure-sdk-for-go/sdk/azcore@v1.14.0)
instead of the deprecated autorest. It's part of [[Epic] Replace
deprecated REST and auth packages in Azure Native
#3576](https://github.com/pulumi/home/issues/3576). The initial commit
is by @mikhailshilkov, thanks!

The high-level view of the changes is:
- Implement a new `AzureClient` that uses azcore. The `AzureClient`
interface remains unchanged.
- azcore implements polling for long-running requests internally, using
the `Location` header. That isn't always a qualified URL, as I've
observed, so there's new code to make the `Location` qualified.
- azcore wants to authenticate via the new azidentity package which
we're not using yet. We do already have a bridge, however, that gets
tokens from the current auth stack and serves them via azidentity's
interface. We pass that bridge into the new Azure client.
- Using the new client is a small change in provider.go. Since
regressions of this change could be very impactful and blocking, there's
an escape hatch `PULUMI_USE_AUTOREST` environment variable to revert to
the old client.
- "**Fix older bug in CanCreate**" and "**TokenCredential bridge: use
correct environment**" are fixes for already existing bugs that I've
found.

Integration testing is essentially provided by the existing test suite
since all tests now use the new Azure client.

---------

Co-authored-by: Mikhail Shilkov <github@mikhail.io>