New scheduled workflow to test using the azcore backend nightly (#3591)
Since we chose an opt-in approach for the time being, the new azcore
client from #3536 is not tested by regular CI workflows. This PR adds a
nightly workflow that tests it. It's a stripped-down copy of
`master.yaml`.

We can then easily add the opt-in for azidentity auth as well.

[Here's a recent
run.](https://github.com/pulumi/pulumi-azure-native/actions/runs/11052940342/job/30707287412)
The only failures are OIDC tests, which should pass on `master` because
there's an identity defined on our Service Principal for the master
branch.