Bugfix: remove accidental commit of test code in Diff (#3633)
It seems like I accidentally left some debugging code in #3554 that
breaks `Diff`. I don't understand how this only comes up now in triage -
wouldn't this mess up _all_ the provider's diffs??

Also a good example for the importance of test coverage. We do have
tests for the diff algorithm but not for the providers overall `Diff`
method.

Fixes #3632