Enable full Read of SqlVirtualMachine by enabling additional GET query params (#3639)
The `sqlvirtualmachine:SqlVirtualMachine` resource is special in that it
has a query parameter `$expand` for GET requests. If `$expand` is not
set, only a subset of properties is returned. That's a problem for
importing this resource since the generated state and code are
incomplete. By setting `$expand`, we can get all properties.

To that end, this PR adds a new field `ReadQueryParams ` to resources in
metadata, containing params and their values that should be set for all
read requests.

A new data structure in gen/readParams.go records these cases so
schema/metadata generation can set the field.

Can be reviewed commit by commit:
- **New metadata field to record URL query params to be used for read
requests**
- **Use new metadata field of URL query params for read requests**
- **Document new customizability in playbook**

Fixes #3581.

Unfortunately, I wasn't able to test the fix e2e, since
SqlVirtualMachine requires a license.