Retry Conflicts for AnotherOperationInProgress
The new azclient allows us to pass a callback function to decide if we want to retry a request. The default behaviour is to only retry a specific set of status codes. This extends that behaviour to also check Conflict (409) and retry if the "error.code" field is "AnotherOperationInProgress".

We're avoiding retrying all conflict status codes because some won't be fixed with a retry, so we're targeting the specific case instead.

Tested this both manually stepping through the code with a debugger and also created an integration test which previously failed but now passes when configured to use the new azclient instead of autorest client.

To access this improved behaviour right now, set the environment variable `PULUMI_USE_AUTOREST=false`. This will become the default in the future, but is currently in testing behind a feature flag.

Fixes #903