Retry Conflicts for AnotherOperationInProgress (#3653)
The new azclient allows us to pass a callback function to decide if we
want to retry a request. The default behaviour is to only retry a
specific set of status codes. This extends that behaviour to also check
Conflict (409) and retry if the "error.code" field is
"AnotherOperationInProgress".

We're avoiding retrying all conflict status codes because some won't be
fixed with a retry, so we're targeting the specific case instead.

Tested this both manually stepping through the code with a debugger and
also created an integration test which previously failed but now passes
when configured to use the new azclient instead of autorest client.

To access this improved behaviour right now, set the environment
variable `PULUMI_ENABLE_AZCORE_BACKEND=true`. This will become the
default in the future, but is currently in testing behind a feature
flag.

This should address #903 which we will close once we've defaulted to
using the new azcore implementation:
- https://github.com/pulumi/pulumi-azure-native/pull/3654