Remove aliases for lower-cased types introduced for v2 (#3749)
## Description

Removes the generation of aliases for the v1 lower-cased types, when the
major version is 3 or up.

The generated code is unchanged for version 2, which is why this PR
doesn't contain a bunch of generated code. It is expected the generated
code will change when the `PROVIDER_VERSION` (i.e. the `majorVersion`
within the generator) reaches 3.x.

## Testing
A new unit test was added: gen_aliases_test.go

Closes #2411