Custom delete for dbforpostgresql:Configuration to reset to the correct default (#3752)
Resolves #3476.

The `dbforpostgresql:Configuration` resource cannot actually be deleted,
so we reset it to its default value instead. However, there are many
configuration parameters with different defaults. On delete, we first
GET this configuration to read the correct default value, then set it.
This is the same approach that `az` takes in `az postgres
flexible-server parameter set`.