[v3] Remove __inputs in state (#3772)
This PR implements the next stage of removing the `__inputs` from state.
In this stage, the provider has "v2" and "v3" behavior. In "v2" mode,
the behavior is unchanged. In "v3" mode, `Configure` asserts that the
engine is passing the old inputs, and doesn't write `__inputs` to state.
The fallback code that reads the `__inputs` from state is effectively
inoperative in the "v3" mode.

Added a line item to this ticket for the final phase which is to remove
the "v2" code:
https://github.com/pulumi/pulumi-azure-native/issues/3754

Closes #2686