Implement autonaming configuration protocol (#3779)
This PR implement the Azure Native part of
https://github.com/pulumi/pulumi/issues/1518. See
https://github.com/pulumi/pulumi/discussions/17592 for the full design.

In short, https://github.com/pulumi/pulumi/pull/17810 introduced the
protobuf changes required for the provider-side implementation. With
those, a provider can:

- Declare that it supports autonaming configurations with a response
flag in Configure
- Accept two extra properties in CheckRequest: a proposed name and a
mode to apply it

This PR applies those parameters to auto-name calculation if they are
specified:

- In Disabled mode, we don't calculate auto-names and let the resource
validation fail if an explicit name is not specified
- In Enforce mode, we use whatever ProposedName is supplied by the
engine
- In Propose mode, almost the same as Enforce with exception of
calculating a UUID for UUID fields and ignoring a proposed name

Added unit tests to cover most autonaming cases (old or new). Also,
added an end-to-end test in YAML.

Resolves https://github.com/pulumi/pulumi-azure-native/issues/3777