Custom support for File Share Protected Items (#3809)
Resolves #1420

The
[recoveryservices.ProtectedItem](https://www.pulumi.com/registry/packages/azure-native/api-docs/recoveryservices/protecteditem/)
represents a resource that's backed up via Azure Backup Service (which
confusingly has the namespace Microsoft.RecoveryServices). ProtectedItem
is a sort of union type with one member per Azure resource that can be
protected, e.g., AzureVmWorkloadProtectedItem.

One of these types, AzureFileshareProtectedItem, is special in that it
requires a file share name as input that's _not_ the name of the file
share as given by the user and as shown by the portal, but an internal
"system name". Therefore, attempting to use AzureFileshareProtectedItem
in the standard way is not possible, because for file share "Foo",
`"protectedItemName": "Foo"` will not work.

Instead, we need to look up the system name for the user-visible name
"Foo" first, and use it instead.