[v3] Detect resource path conflicts and disambiguate (#3817)
Partly fixes #2495, #2753.

Previously, schema and metadata generation of this provider assumed that
a type token like `azure-native:netapp:Pool` uniquely identifies a
resource. In the Azure spec, that's not always the case across API
versions.

For instance, as shown by #2753, resources such as
`PrivateEndpointConnection ` can point to the /servers/ API in one API
version but to the newer /flexibleServers/ API, which is a completely
different product underneath, in another version.

This PR does two things:
1. Detect such conflicts. So far, it only prints a warning when it finds
one, but when all such conflicts are investigated and fixed it should
error.
2. Disambiguate several known instances of such conflicts. In the above
example of `PrivateEndpointConnection`, we rename the resource to
`SingleServerPrivateEndpointConnection` when it points to the /servers/
API.

Any change to the generated schema is gated behind a "version >= 3"
check to avoid breaking changes in v2.