Unify the current version
Previously, we had the version both embedded via ldflags into `version.Version` but also passed in to the gen program as a CLI argument.
The version should _always_ be embedded during a proper build, but might be missing for when debugging via an IDE.

1. Make everything work without having to pass the version as a CLI argument.
2. Assert that the version is set by exactly one of the CLI argument or the ldflags for consistency.
3. Use the parsed semver.Version object everywhere after initialization.

This removes passing the version by CLI completely in the makefile.