Custom resource for Azure PIM role management policies (#3663)
## About

This PR adds support for [Role Management
Policies](https://learn.microsoft.com/en-us/rest/api/authorization/privileged-role-policy-rest-sample),
part of [Privileged Identity Management
(PIM)](https://learn.microsoft.com/en-us/entra/id-governance/privileged-identity-management/pim-configure)
in the Microsoft.Authentication namespace. It's one part of #2455.

Note that this is about the ARM part of PIM; there's also a Microsoft
Graph API part which is not covered by this provider.

This resource wasn't automatically included because it supports only GET
and PATCH. The policies are singletons that cannot be created or
deleted, only modified via PATCH.

## Implementation

Role Management Policies essentially consist of a name which is actually
a GUID, and a list of ~20 rules.

Using our existing singleton support `defaults.GetDefaultResourceState`
was tricky because
1. there are many policies with many rules, for a total of ~300k lines
of JSON for a subscription scope, and there are more scopes, plus
2. I believe the defaults can vary per scope and possibly also per
customer.

So instead, I've implemented a custom resource that captures the
original state of a policy when it's first "created", i.e., added to
Pulumi state. When a rule or the whole policy is removed from Pulumi, we
look up the original state and re-apply it.

## Testing

The e2e/integration test for this resource is special because using PIM
requires a paid [Entra ID P2
license](https://www.microsoft.com/en-us/security/business/microsoft-entra-pricing).
We have one that you can see
[here](https://admin.microsoft.com/Adminportal/Home?referrer=entra#/licensedetailpage/84a661c4-e949-4bd2-a560-ed7766fcaf2b).