GH-3963: Allow user to specify audience for OIDC authentication (#3970)
This PR is a copy of #3965 by @robcao, to make CI tests run. Original
description:

-----

This pull request is a potential implementation for
https://github.com/pulumi/pulumi-azure-native/issues/3963

It adds a new configuration variable for the auth_azidentity module,
ARM_OIDC_AUDIENCE.

When specified, the value of ARM_OIDC_AUDIENCE will be used when making
a request for a federated token. If not specified, the audience will
default to api://AzureADTokenExchange like currently.

Co-authored-by: Robert Cao <robert.cao@relativity.com>