Support for PIM Role Eligibility Schedules (#3959)
Resolves #2455 - add support for PIM (Privileged Identity Management)
Role Eligibility Schedules.

- [Azure docs](https://learn.microsoft.com/en-us/rest/api/authorization/privileged-role-eligibility-rest-sample)
-
[Spec](https://github.com/Azure/azure-rest-api-specs/tree/main/specification/authorization/resource-manager/Microsoft.Authorization/stable/2020-10-01)

This API has some peculiarities that required writing a custom resource:

1. It's divided into schedules and schedule requests (an endpoint
`/roleEligibilityScheduleRequests`, not HTTP requests). Requests are
created to create and also delete schedules; there's no DELETE endpoint.
Requests are short-lived but can require approval.

2. Because of this duality where two types are used, schedule and
schedule request, I extended custom resources with the ability to derive
their schema from type A in the spec but define a different resource
name B. We want the resource to be named `...Schedule` but need the
`...ScheduleRequest` types.

3. While a schedule request is pending, a `/cancel` endpoint is used to
delete the request.

4. Both schedules and schedule requests cannot be modified; new ones
must be created instead. This isn't properly annotated in the spec,
though. The SDKs are correct in that they don't offer update operations.

I tried to comment the code pretty well, please refer to the Godoc
comments for more in-depth information.