Re-compress the Python Wheel package with bzip2 to shave off 7 MB. (#3982)
The provider is currently blocked from releasing the Python SDK because
the size of the Wheel package reached 100 MB, see #3977.

While we're waiting on PyPi to increase our limit (if they do), I looked
into better compression. Turns out that the package is currently
compressed via zip's deflate method, while the better bzip2 would also
be valid.

I first tried to use bzip2 directly when building the package, but
failed due to https://github.com/pypa/setuptools/issues/2491. This PR
instead builds the Wheel as before, then unzips it and re-zips it with
bzip2.

A [test
run](https://github.com/pulumi/pulumi-azure-native/actions/runs/13507283628/job/37742131092)
shows that it works. The package is 93.8 MB and is uploaded. It fails
after upload with _The use of local versions in
`Version(2.89.0a1740429149+ded80e4)` is not allowed_ but that's because
[provider-version-action adds the local commit when it's a workflow run
that's not on the main
branch](https://github.com/pulumi/provider-version-action/tree/v1/?tab=readme-ov-file#scenarios).