Fix for async deletion of non-existent resource (#3980)
This PR fixes an error that occurs in the deletion of a (non-existent)
Azure resource. The issue impacts resources that have an async-style
DELETE call (as is the case for Subnet, see
[schema](https://github.com/Azure/azure-rest-api-specs/blob/1b038fc40563bdafb343a71cd5f7d3dee6bc4b5b/specification/network/resource-manager/Microsoft.Network/stable/2019-02-01/virtualNetwork.json#L396-L399)).
The technical detail is that the 404 may come in the initial response as
opposed to during polling, and that wasn't handled correctly..

The user-facing error is:
```
  azure-native:network:Subnet (subnet):
    error: the operation failed or was cancelled
```

The call to `runtime.NewPoller` should not be made with a failed
response, otherwise we hit a precondition within it:
```go
	// this is a back-stop in case the swagger is incorrect (i.e. missing one or more status codes for success).
	// ideally the codegen should return an error if the initial response failed and not even create a poller.
	if !poller.StatusCodeValid(resp) {
		return nil, errors.New("the operation failed or was cancelled")
	}
```

To fix, I encapsulated the request-to-final-response logic into an
anonymous function, to be able to uniformly check for the 404 result.
And I ensure that the initial response was successful before making a
poller.

This PR relaxes the logic for detecting a 404 during resource deletion,
to not care about the specific error code. I find that at least two
codes are possible: `ResourceNotFound` and `ResourceGroupNotFound`. Note
that the sync path did not check the error code anyway and a uniform
solution is desirable.

TODO:

- [x] new unit tests
- [x] manual testing for various types of resources (w/ sync and async
delete operations).

Closes #3978
See also: https://github.com/pulumi/pulumi-azure-native/pull/3783