Generate a default API versions only schema and use it for sdkgen in version 3+ (#4016)
For v3 of the provider, we want to generate SDKs only for the default API versions. However, to enable parameterization to generate SDKs for any API version at runtime, we need the full schema and metadata. (Other approaches like downloading on demand are possible but not implemented at this time.) This PR adds a build step of generating a default-versions-only schema that's then used to generate the SDKs. The full schema is embedded into the provider as before.

A branch with this PR but regenerated for `PROVIDER_VERSION=3.0.0` is
[here](https://github.com/pulumi/pulumi-azure-native/tree/tkappler/v3-default-only-genv3).

Resolves #3422