Add Parameterize to generate an SDK for a specific API version (#4010)
```
> PULUMI_DEBUG_PROVIDERS="azure-native:60383" pulumi package add azure-native storage v20240101
Successfully generated a Nodejs SDK for the azure-native_storage_v20240101 package at /Users/tkappler/pulumi/repro/azn-param/sdks/azure-native_storage_v20240101

> head -3 index.ts
import * as resources from "@pulumi/azure-native/resources";
import * as storage_v20240101 from "@pulumi/azure-native_storage_v20240101";

> rg _storage package.json
10:        "@pulumi/azure-native_storage_v20240101": "file:sdks/azure-native_storage_v20240101",

> PULUMI_DEBUG_PROVIDERS="azure-native:60383" pulumi up
...
 +   └─ azure-native_storage_v20240101:storage/v20240101:StorageAccount  sa             created (27s) 
```

I picked an approach that's, I hope, pragmatic rather than elegant, in
that I assume we can keep the full schema and metadata available for the
provider and generate dedicated versioned schemas simply by picking the
required parts from the full schema. This avoids having to publish and
download or clone data at generation time. A more dynamic approach is
equally possible, of course, and could be swapped in without changing
the interface.

To understand the PR you should be familiar with [these
docs](https://pulumi-developer-docs.readthedocs.io/latest/docs/architecture/providers/parameterized.html).

The input and behavior of Parameterize is the same in the initial "args"
call of Parameterize and in the subsequent "serialized parameters"
calls: a pair of (Azure module, API version). The existing full schema
and metadata are narrowed down to members belonging to the given module
at the given version.

_Implementation note on `resources.PartialAzureAPIMetadata` and
`resources.APIMetadata`_: the provider used to store metadata in a
`PartialAzureAPIMetadata`. During Parameterize, we couldn't replace this
metadata with the new one which was of type `AzureAPIMetadata`. Using
the new `APIMetadata` we can accomodate both because it uses `MapLike`
which can be a full or a partial map. It turned out that this change
also simplified some other code using these types.

Related to #2467