In v3, generate default versions only SDK for Go, too (#4036)
This is a follow-up to #4016 where I had forgotten that Go SDK
generation works differently than the other languages. Fortunately, the
code already accepts the required arguments, we just need to pass them.

This PR has no effect on the SDK in v2.x, only in v3.x.

I wonder if we should remove the "default" schema location altogether
and just make the argument mandatory.