Update path conflict resolution for newer API versions (#4053)
Turns out that some of them are not necessary anymore and removing them
simplifies the resource names. We know they're not necessary because of
the check in schema.go that fails the build when there are path
conflicts. This PR also makes this check required for non-release builds
now that all de-duplications are in place.