Add a message about other API versions to v3 registry API docs (#4056)
Resolves #4054

This results in no changes for v2, and looks like in this screenshot in v3.

![Screenshot 2025-03-25 at 10 48
43](https://github.com/user-attachments/assets/d6a6c542-0a41-49c8-99ad-7cae1f4b96a1)

Details to be discussed, like the exact wording and whether there should
be a paragraph break before "_These versions are not included ..._".

The biggest issue I see is hard-coding the registry link in provider
code, but I couldn't think of a better way.