run e2e tests in "v3" mode (#4182)
This PR updates the e2e tests to use "v3" mode, by setting the
`version.Version` global to the value of `PROVIDER_VERSION`. This
variable is used by the provider to switch between v2 and v3 behavior,
and clearly the e2e tests should be using v3.

Note that the e2e tests do not contribute to code coverage.

Closes #4181